Fix bad explanation and rename module
